content-processing/mlcp/src/requirements.txt
boto3==1.28.68
gremlinpython==3.7.0
matplotlib==3.8.0
networkx==3.2
numpy==1.24.3
opencv_python==4.7.0.72
pandas==2.1.1
pdf2image==1.16.3
Pillow==10.1.0
psutil==5.9.5
pytesseract==0.3.10
spacy==3.7.2
